Clarify what is being debounced
The use of `Debounce` in the description leads to confusion as to what is being debounced. This double usage makes it unclear whether we are debouncing the time idle or the time active. As such, this pull request clarifies by explaining that the debounce is simply the minimum idle time for the user before firing the action.